Update of the Reusable Table Output Component
